# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EVA API",
        "contact": {
            "email": "gurkalov.dev@gmail.com"
        },
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/v1/routers": {
            "get": {
                "tags": [
                    "Router"
                ],
                "description": "Returns all routers",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "maximum number of results to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Router response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Router"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Router"
                ],
                "description": "Creates a new router.",
                "operationId": "createRouter",
                "requestBody": {
                    "description": "Router to add.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewRouter"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Router response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Router"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/routers/{id}": {
            "get": {
                "tags": [
                    "Router"
                ],
                "description": "Returns a router based on a single ID",
                "operationId": "getRouter",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of router to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Router response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Router"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Router"
                ],
                "description": "Update a router based on a single ID.",
                "operationId": "updateRouter",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of router to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Router to update.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewRouter"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Router response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Router"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Router"
                ],
                "description": "Deletes a single router based on the ID.",
                "operationId": "deleteRouter",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of router to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Router deleted"
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorModel": {
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "NewRouter": {
                "required": [
                    "name"
                ],
                "properties": {
                    "ssid": {
                        "type": "string",
                        "format": "string"
                    },
                    "bssid": {
                        "type": "string",
                        "format": "string"
                    }
                },
                "type": "object"
            },
            "Router": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/NewRouter"
                    },
                    {
                        "required": [
                            "id"
                        ],
                        "properties": {
                            "id": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "ssid": {
                                "type": "string",
                                "format": "string"
                            },
                            "bssid": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        }
    }
}